--- src/@types/enums/index.ts
@@ -1,7 +1,0 @@
-const enum Status {
-  LOADING = 'loading',
-  SUCCESS = 'success',
-  ERROR = 'error',
-}
-
-export { Status };
