+++ src/components/CartItem.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { addItem, minusItem, removeItem, TCartItem } from '../redux/slices/cartSlice';
+import { addItem, minusItem, removeItem } from '../redux/cart/slice';
+import { TCartItem } from '../redux/cart/types';
 
 const CartItem: React.FC<TCartItem> = ({ id, title, type, size, price, quantity, imageUrl }) => {
   const dispatch = useDispatch();
