+++ src/redux/cart/types.ts
@@ -1,0 +1,16 @@
+type TCartItem = {
+  id: string;
+  title: string;
+  price: number;
+  imageUrl: string;
+  type: string;
+  size: number;
+  quantity: number;
+};
+
+interface ICartSliceState {
+  totalPrice: number;
+  items: TCartItem[];
+}
+
+export type { TCartItem, ICartSliceState };
