+++ src/redux/filter/selectrors.ts
@@ -1,0 +1,6 @@
+import { RootState } from '../store';
+
+const selectSort = (state: RootState) => state.filter.sort;
+const selectFilter = (state: RootState) => state.filter;
+
+export { selectSort, selectFilter };
