+++ src/redux/store.ts
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { useDispatch } from 'react-redux';
-import filter from './slices/filterSlice';
-import cart from './slices/cartSlice';
-import pizza from './slices/pizzaSlice';
+import filter from './filter/slice';
+import cart from './cart/slice';
+import pizza from './pizza/slice';
 
 export const store = configureStore({
   reducer: { filter, cart, pizza },
